fix(terraform-provider): renamed provider.ts to providers.ts
